common: Move method to decode an address without HRP verification
